requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.10.1
mysqlclient==2.1.1
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
